--- src/action_types/actionTypes.ts
@@ -1,11 +1,0 @@
-export const ASYNC_ACTION_START = 'common/ASYNC_ACTION_START';
-export const ASYNC_ACTION_END = 'common/ASYNC_ACTION_END';
-export const CONFIRM_ACTION = 'common/CONFIRM_ACTION';
-export const CONFIRM_ACTION_CANCEL = 'common/CONFIRM_ACTION_CANCEL';
-
-export const LOAD_STUDENTS_STATISTICS = 'student/LOAD_STUDENTS_STATISTICS';
-export const LOAD_STUDENTS = 'student/LOAD_STUDENTS';
-export const LOAD_STUDENT = 'student/LOAD_STUDENT';
-export const CREATE_STUDENT = 'student/CREATE_STUDENT';
-export const UPDATE_STUDENT = 'student/UPDATE_STUDENT';
-export const DELETE_STUDENT = 'student/DELETE_STUDENT';
