+++ src/action_types/studentsActionTypes.ts
@@ -1,0 +1,6 @@
+export const LOAD_STUDENTS_STATISTICS = 'students/LOAD_STUDENTS_STATISTICS';
+export const LOAD_STUDENTS = 'students/LOAD_STUDENTS';
+export const LOAD_STUDENT = 'students/LOAD_STUDENT';
+export const CREATE_STUDENT = 'students/CREATE_STUDENT';
+export const UPDATE_STUDENT = 'students/UPDATE_STUDENT';
+export const DELETE_STUDENT = 'students/DELETE_STUDENT';
